requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.3
python-http-client==3.3.7
pytz==2022.2.1
requests==2.28.1
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.1.0
urllib3==1.26.12
whitenoise==6.2.0
